% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_add_ribbon.R
\name{.add_ribbon}
\alias{.add_ribbon}
\title{Generate a Fixed-Width Decorative Ribbon}
\usage{
.add_ribbon(msg, char = "=", total_width = 80)
}
\arguments{
\item{msg}{Character. The text message to be displayed.}

\item{char}{Character. The character used to draw the ribbon. Default is \code{"="}.}

\item{total_width}{Integer. The target total width of the string. Default is \code{80}.}
}
\value{
A character string of length \code{total_width} (or \code{nchar(msg)} if
\code{msg} exceeds \code{total_width}).
}
\description{
Creates a string consisting of a message followed by a repeated character
ribbon, padding the total length to a specified width. This is used for
formatting log headers and visual separators in the console output.
}
\examples{
# Internal usage:
# add_ribbon("== [ STARTING PROCESS ] ", char = "=", total_width = 80)

}
