% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_log_helpers.R
\name{log_cotan_execution}
\alias{log_cotan_execution}
\title{Handle the start and end of a COTAN execution with native output tracking}
\usage{
log_cotan_execution(func_name, ..., is_complete = FALSE, level = 1L)
}
\description{
Handle the start and end of a COTAN execution with native output tracking
}
