% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_log_helpers.R
\name{log_header}
\alias{log_header}
\title{Print a formatted header with dynamic clean tree}
\usage{
log_header(msg, is_complete = FALSE, level = 3L)
}
\description{
Print a formatted header with dynamic clean tree
}
