% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_log_helpers.R
\name{log_matrix_stats}
\alias{log_matrix_stats}
\title{Log Matrix Filtering Statistics}
\usage{
log_matrix_stats(
  cells_before,
  cells_after = NULL,
  features_before = NULL,
  features_after = NULL,
  level = 1L
)
}
\description{
Log Matrix Filtering Statistics
}
