% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{reset_log_state}
\alias{reset_log_state}
\title{Reset the logging state}
\usage{
reset_log_state()
}
\value{
No return value.
}
\description{
Resets the internal indentation counter to zero.
It is highly recommended to call this function at the very beginning
of your main pipeline script to ensure a clean state, or after a manual interruption.
}
