% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_save_obj.R
\name{save_object}
\alias{save_object}
\title{Save an R Object to a Specific Directory}
\usage{
save_object(obj, out_dir, file_name)
}
\arguments{
\item{obj}{The R object to be saved (e.g., a Seurat or COTAN object).}

\item{out_dir}{Character. The destination directory path.}

\item{file_name}{Character. The name of the file. It is highly recommended
to include the \code{.rds} extension (e.g., \code{"cotan_filtered.rds"}).}
}
\value{
Invisible character string containing the absolute or relative
full path to the saved file.
}
\description{
A utility helper function to safely store R objects (such as
\code{COTAN} or \code{Seurat} instances) to disk as \code{.rds} files. It automatically
checks if the target directory exists and creates it recursively if necessary.
}
